quarantining.html dummy response all set. client side is done.
update: quarantining.html dummy response all set. client side is done. Now working on backend receiving(I know we don't have to do that now but for the sake of the next one). I'll push it rn remember to use port 3000. I also moved all the html and css to a folder named "public" so express is easier to deal with it